+++ src/components/ShareExperience/questionCreators.js
@@ -68,7 +68,7 @@ import {
 import { QUESTION_TYPE } from '../common/FormBuilder/QuestionBuilder';
 import { salaryHint } from 'utils/formUtils';
 import { useTotalCount } from 'hooks/useCount';
-import CompanyAutoCompleteItem from './CompanyAutoCompleteItem';
+import AutoCompleteItem from './AutoCompleteItem';
 
 export const createCompanyQuestion = ({ header }) => ({
   title: '公司名稱',
@@ -89,7 +89,7 @@ export const createCompanyQuestion = ({ header }) => ({
     fetchSearchCompany({ companyName: value }).then(
       map(({ name, businessNumber }) => ({
         label: (
-          <CompanyAutoCompleteItem
+          <AutoCompleteItem
             pageType={PAGE_TYPE.COMPANY}
             name={name}
             businessNumber={businessNumber}
